refactor transform mapping module
